moved cache cgheck to activity viewmodel
