+++ app/app/src/main/res/layout/item_list_link.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
+        <import type="android.view.View" />
 
         <variable
             name="item"
@@ -47,6 +48,18 @@
                 android:text="@{item.type, default=`Torrent`}"
                 app:layout_constraintBottom_toTopOf="@id/link"
                 app:layout_constraintTop_toTopOf="parent"
+                />
+
+            <TextView
+                android:id="@+id/cached"
+                style="@style/TextAppearance.UnchainedTheme.Overline"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="5dp"
+                app:layout_constraintStart_toEndOf="@id/type"
+                android:text="@string/cached"
+                app:layout_constraintTop_toTopOf="parent"
+                android:visibility="@{item.cached == false ? View.GONE : View.VISIBLE}"
                 />
 
             <TextView
